fix navbar home label and navbar paths, move navbar from Home.jsx to App.jsx
