+++ src/App.js
@@ -2,6 +2,8 @@ import React, { createContext } from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import { Home, Header } from './components/';
 import { useGeolocation } from './hooks&aux/useGeolocation';
+import { Navbar } from './components/navbar/Navbar.jsx'
+
 
 export const LocationContext = createContext()
 
@@ -16,6 +18,7 @@ export const App = () => {
           <Route path='/' element={<Home />} />
           <Route path='*' element={<h1>404 Not Found</h1>} />
         </Routes>
+        <Navbar />
       </LocationContext.Provider>
     </Router>
   );
