+++ src/components/Home.jsx
@@ -5,7 +5,7 @@ import { getProducts, getRestaurants } from '../services'
 import styles from './home.module.css'
 import { useInfiniteScroll } from '../hooks&aux/useInfiniteScroll'
 import { filterResults } from '../hooks&aux/filterResults';
-import { Navbar } from './navbar/Navbar';
+
 
 export const Home = () => {
 
@@ -44,7 +44,7 @@ export const Home = () => {
                 clearResults={clearResults} />
             {loading ? <> <Skeleton /> <Skeleton /> </> : <Products id='productList' arrayProduct={products} search={restaurants} />}
             {loadSkeleton && <><Skeleton /></>}
-            <Navbar />
+            
         </div>
     )
 }
