+++ src/components/navButtons/Buttons.jsx
@@ -4,17 +4,21 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 
 
 export const Buttons = ({ name, icon, pressed, setPressed }) => {
-  
+
   const [isActive, setIsActive] = useState(true);
 
-  const homeTitle = name==="home"?"home":null;
+  const homeTitle= ()=>{
+    if (pressed === name && name === "home") {
+      return "home";
+    }
+  }
 
   const handleIsActive = () => {
 
     if (!isActive) {
       setIsActive(true);
       setPressed(name);
-    
+
     }
     //  else {
     //   setIsActive(false);
@@ -24,7 +28,14 @@ export const Buttons = ({ name, icon, pressed, setPressed }) => {
   };
 
   useEffect(() => {
-    pressed === name ? setIsActive(true) : setIsActive(false);
+
+    if (pressed === name) {
+      setIsActive(true);
+
+    } else {
+      setIsActive(false);
+    }
+
 
   }, [name, pressed])
 
@@ -34,26 +45,26 @@ export const Buttons = ({ name, icon, pressed, setPressed }) => {
 
     color: "#31B9CC",
     height: "16px",
-    
+
 
   };
 
   const offActive = {
     color: "#C6C7CA",
     height: "16px",
-    
+
   };
 
   const onActiveShadow = {
     color: "#31B9CC",
     backgroundColor: "#F1F2F6"
-    
+
   };
 
   const offActiveShadow = {
     color: "#C6C7CA",
     backgroundColor: "#FFFFFF"
-    
+
   };
 
 
@@ -63,12 +74,8 @@ export const Buttons = ({ name, icon, pressed, setPressed }) => {
     <button className="buttonNavBar" onClick={handleIsActive} style={isActive ? onActiveShadow : offActiveShadow}>
 
       <FontAwesomeIcon className="iconNavBar" icon={icon} style={isActive ? onActive : offActive} />
-      <p>{ homeTitle }</p>
+      <p>{homeTitle()}</p>
 
     </button>
   );
 };
-
-
-
-//navlink
